This is a giant commit.  This is code cleanup in simbot.pl, implementation of all the events I had planned, and splitting out of undernet channel services login into a plugin along with a model for choosing a particular (any only one) services plugin if there are many available.
I would like someone to write a ChanServ/NickServ plugin. :)

I want to thank the staff at Undernet for causing a timely server split,
allowing me to give the Undernet service plugin a few extra tests that I
wasn't certain I'd get in tonight.

If you turn on $verbose = 4 to get the "SPAM" output, you'll see that I've
added a lot of new information to the output, much of which may come in
handy if you're trying to debug plugins and event firing.

All in all, a hefty change and two items out from the todo plugin. Four
more items and then we'll enter a brief "beta" phase, followed by a 6.0
release!  Woot! :)