This commit touches just about every file.  I implemented send_message, send_action, and send_notice to replace plugins and the main script calling POE's kernel directly and revised all plugins to use the new functions.  This allowed me to fix recap to log both "seen" and "sent" text together.
I meanwhile fixed some recap logic that would cause incorrect ordering
of history items (the bot's reply would always come before whatever it
was replying to), as well as a typo or two, and lots of tabbing and
trailing whitespace stuff.

I also made the weather plugin not die on a failure to load its cache.
This is a failure we can live with if it happens.

I may or may not have done something else.  I probably won't know until
I look at the diff later. :)