We check for orphaned words at startup now.  This means any word that doesn't point to any other words in one or both directions will be removed.  Words like this could exist if you edited the database manually.
If the bot ran into one of these words, the only problem would generally be
that it would just stop building a sentence right there.  This used to cause
infinite loops, but I fixed that long ago.  All this really does now is reduce
database bloat and save a tiny bit of memory.

I think it's a good move though.