This commit does almost nothing at all.  There's some whitespace cleanup, and a bunch of new comments.  I also removed some code that was commented out because the superceding code appears to be working at least as well.
The _one_ and _only_ thing that I have changed is that the bot will properly
recognize incorrect % commands when they are typed in all caps.  Before the
command was required to have at least one lowercase letter, and I figure that
if you're going to accidentally capitalize letters in a command, you've
probably hit caps lock and that's going to mean you won't get an error message.
And everyone loves error messages. :)