This is another great idea.  I'm really interested in natural language processing, so this project was fun, and I intend to keep it maintained after the class that this was written for ends.
So what is this?

This is nlp.pl for CS451 (Artificial Intelligence)....

It adds natural language support, via a plugin, to simbot.  It also
modifies the core script to allow the necessary callbacks and data
to be provided by plugins that want to use this functionality, if the
plugin happens to be loaded.

I also threw in a snippet of code that eats IRC log looking junk from
the beginning of lines before recording them, which should keep
timestamps and usernames out of the database.