Recommitting (I accidentally deleted too many revisions)....
SimBot now has a %snooze command which turns on snooze mode.  Snooze mode
prevents simbot from learning or responding convesationally.  Plugins will
need to look at $SimBot::snooze if they would like to try to be quiet when
snooze is on.  Commands still get processed, and plugins like recap were
not modified, because it seems correct to still collect recap data.  Info
might want to stop learning or responding outside of %info, but I'll leave
that up to 404 to decide.