Apparently I never realized PoCo::IRC sends the channel as a reference to an array.  I didn't realize this because I've only ever given the value back to PoCo::IRC and it never seemed to care.  At some point it will matter more to me, when I decide to make SimBot support two or more channels.
This is a beginning attempt at implementing irc command overrides.  This
means a services plugin can replace the usual 'kick' call with its own, for
example, if you wanted your IRC network's channel services to take care of the
kick.  I also modified the undernet plugin to implement kick so I could test
it.  There is lots of room for improvement, but it works for now, and I am the
only one with the exclusive Kick SteveB plugin used to test the code. :)