Added SimBot::Util module which is now used by literally everything.  It's nice because it allows the plugins to use in their namespace several handy utility functions, while also allowing us to truly hide functions and variables that plugins shouldn't be accessing directly.
This module can also be used by scripts in tools/ if they should want to use
any utility functions we have created.  It is a completely independent perl
module.

I will likely be adding SimBot::Plugin and attempting to migrate the plugin
functions and data into that file if possible as my next step.  The other thing
I'd like to do fairly soon is get the main database into SQLite.  We'll see... :)